requirements.txt
flask
flask_wtf
wtforms
flask_login
pillow
flask_mail
flask_sqlalchemy
flask_bcrypt
datetime
itsdangerous